training/requirements.txt
# Treinamento LoRA
unsloth>=2024.11
trl>=0.12
transformers>=4.46
peft>=0.13
datasets>=3.1
torch>=2.5
bitsandbytes>=0.44

# Database
psycopg2-binary>=2.9

# HTTP (deploy vLLM)
requests>=2.32

# Utils
numpy>=2.1
scikit-learn>=1.5
